refactor (all): first attempt of component principle
