+++ src/main.py
@@ -1,8 +1,8 @@
-from console_io import ConsoleIO
-from game_interface import GameInterface
-from game_board import GameBoard
-from game_engine import GameEngine
-from game_policy import GamePolicy
+from game_interface.console_io import ConsoleIO
+from game_interface.game_interface import GameInterface
+from game_play.game_board import GameBoard
+from game_play.game_engine import GameEngine
+from game_play.game_policy import GamePolicy
 
 io = ConsoleIO()
 game_interface = GameInterface(io)
